+++ Content/Items/Weapons/Ranged/Desperado/DesperadoShrapnel.cs
@@ -1,5 +1,8 @@
+using Microsoft.Xna.Framework;
 using Terraria;
+using Terraria.Audio;
 using Terraria.DataStructures;
+using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace SuperfluityTwo.Content.Items.Weapons.Ranged.Desperado
@@ -18,14 +21,39 @@ namespace SuperfluityTwo.Content.Items.Weapons.Ranged.Desperado
 			Projectile.tileCollide = true;
 			Projectile.usesLocalNPCImmunity = true;
 			Projectile.localNPCHitCooldown = -1;
-			Projectile.ArmorPenetration = 18;
 			Projectile.DamageType = DamageClass.Ranged;
 		}
 
-        public override bool PreAI()
+        public override void OnSpawn(IEntitySource source)
         {
-			Projectile.rotation = Projectile.velocity.ToRotation();
-            return true;
+			Dust.NewDustPerfect(
+				Projectile.Center,
+				DustID.GoldFlame,
+				Projectile.velocity * 0.1f
+			);
         }
+		
+        public override void ModifyHitNPC(NPC target, ref NPC.HitModifiers modifiers)
+        {
+			modifiers.FinalDamage /= 2;
+        }
+
+        public override void ModifyHitPlayer(Player target, ref Player.HurtModifiers modifiers)
+        {
+			modifiers.FinalDamage /= 2;
+        }
+
+		public override bool PreAI()
+		{
+			Projectile.rotation = Projectile.velocity.ToRotation();
+			return true;
+		}
+		
+		public override bool OnTileCollide(Vector2 oldVelocity) {
+			Collision.HitTiles(Projectile.position, Projectile.velocity, Projectile.width, Projectile.height);
+			SoundEngine.PlaySound(SoundID.Item10, Projectile.position);
+
+			return true;
+		}
     }
 }